refactor: 상수 추가, 변경, 적용( dropdown, text )
