--- src/constants/sports.ts
@@ -1,5 +1,0 @@
-export const SPORTS_CATEGORY = [
-  { id: 0, text: '축구', value: { sportsCategory: 'SOCCER' } },
-  { id: 1, text: '야구', value: { sportsCategory: 'BASEBALL' } },
-  { id: 2, text: '배드민턴', value: { sportsCategory: 'BADMINTON' } },
-];
